reactors/echo-fastapi/pyproject.toml
[tool.poetry]
authors = ["Brandon Bjelland <brandon@atscale.run>"]
description = ""
name = "echo-fastapi"
packages = [{include = "src"}]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
fastapi = "^0.88.0"
google-cloud-logging = "^3.3.1"
pydantic = "^1.10.2"
python = "^3.10"
uvicorn = "^0.20.0"
google-cloud-deploy = "^1.6.0"
google-cloud-secret-manager = "^2.15.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
pre-commit = "^2.20.0"
pylint = "^2.15.9"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
